+++ base/usr/include/kernel/process.h
@@ -53,7 +53,7 @@ typedef struct image {
 /* Resizable descriptor table */
 typedef struct descriptor_table {
 	fs_node_t ** entries;
-	size_t    *  offsets;
+	uint64_t  *  offsets;
 	int       *  modes;
 	size_t       length;
 	size_t       capacity;
